# api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Chat API",
    "description": "Multi-LLM chat service API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/admin/rate_limit": {
      "post": {
        "operationId": "updateUserRateLimit",
        "summary": "Update a user rate limit",
        "tags": [
          "Admin users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RateLimitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/rateHTTPResponse"
                }
              }
            }
          },
          "default": {
            "description": "API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/session_messages/{sessionUuid}": {
      "get": {
        "operationId": "getAdminSessionMessages",
        "summary": "Get messages for a session",
        "tags": [
          "Admin users"
        ],
        "parameters": [
          {
            "name": "sessionUuid",
            "in": "path",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "items": {
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "uuid": {
                        "type": "string"
                      },
                      "role": {
                        "type": "string"
                      },
                      "content": {
                        "type": "string"
                      },
                      "reasoningContent": {
                        "type": "string"
                      },
                      "model": {
                        "type": "string"
                      },
                      "tokenCount": {
                        "type": "integer"
                      },
                      "userId": {
                        "type": "integer"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    },
                    "additionalProperties": false,
                    "type": "object",
                    "required": [
                      "id",
                      "uuid",
                      "role",
                      "content",
                      "reasoningContent",
                      "model",
                      "tokenCount",
                      "userId",
                      "createdAt",
                      "updatedAt"
                    ]
                  },
                  "type": "array"
                }
              }
            }
          },
          "default": {
            "description": "API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/user_analysis/{email}": {
      "get": {
        "operationId": "getUserAnalysis",
        "summary": "Get user analysis",
        "tags": [
          "Admin users"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAnalysisData"
                }
              }
            }
          },
          "default": {
            "description": "API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/admin/user_session_history/{email}": {
      "get": {
        "operationId": "getUserSessionHistory",
        "summary": "Get user session history",
        "tags": [
          "Admin users"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "maximum": 500,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sessionHistoryPageHTTPResponse"
                }
              }
            }
          },
          "default": {
            "description": "API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/user_stats": {
      "post": {
        "operationId": "getUserStats",
        "summary": "Get user statistics",
        "tags": [
          "Admin users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/pageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/userStatsPageHTTPResponse"
                }
              }
            }
          },
          "default": {
            "description": "API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/users": {
      "put": {
        "operationId": "updateAdminUser",
        "summary": "Update a user",
        "tags": [
          "Admin users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/updateAuthUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/authUserHTTPResponse"
                }
              }
            }
          },
          "default": {
            "description": "API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "RateLimitRequest": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "email": {
            "type": "string"
          },
          "rateLimit": {
            "type": "integer"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "rateLimit"
        ]
      },
...
      "UserAnalysisData": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "userInfo": {
            "properties": {
              "email": {
                "type": "string"
              },
              "totalMessages": {
                "type": "integer"
              },
              "totalTokens": {
                "type": "integer"
              },
              "totalSessions": {
                "type": "integer"
              },
              "messages3Days": {
                "type": "integer"
              },
              "tokens3Days": {
                "type": "integer"
              },
              "rateLimit": {
                "type": "integer"
              }
            },
            "additionalProperties": false,
            "type": "object",
            "required": [
              "email",
              "totalMessages",
              "totalTokens",
              "totalSessions",
              "messages3Days",
              "tokens3Days",
              "rateLimit"
            ]
          },
          "modelUsage": {
            "items": {
              "properties": {
                "model": {
                  "type": "string"
                },
                "messageCount": {
                  "type": "integer"
                },
                "tokenCount": {
                  "type": "integer"
                },
                "percentage": {
                  "type": "number"
                },
                "lastUsed": {
                  "type": "string",
                  "format": "date-time"
                }
              },
              "additionalProperties": false,
              "type": "object",
              "required": [
                "model",
                "messageCount",
                "tokenCount",
                "percentage",
                "lastUsed"
              ]
            },
            "type": "array"
          },
          "recentActivity": {
            "items": {
              "properties": {
                "date": {
                  "type": "string",
                  "format": "date-time"
                },
                "messages": {
                  "type": "integer"
                },
                "tokens": {
                  "type": "integer"
                },
                "sessions": {
                  "type": "integer"
                }
              },
              "additionalProperties": false,
              "type": "object",
              "required": [
                "date",
                "messages",
                "tokens",
                "sessions"
              ]
            },
            "type": "array"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "userInfo",
          "modelUsage",
          "recentActivity"
        ]
      },
...
      "authUserHTTPResponse": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "id": {
            "type": "integer"
          },
          "lastLogin": {
            "type": "string",
            "format": "date-time"
          },
          "isSuperuser": {
            "type": "boolean"
          },
          "username": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "isStaff": {
            "type": "boolean"
          },
          "isActive": {
            "type": "boolean"
          },
          "dateJoined": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "lastLogin",
          "isSuperuser",
          "username",
          "firstName",
          "lastName",
          "email",
          "isStaff",
          "isActive",
          "dateJoined"
        ]
      },
...
      "pageRequest": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "page": {
            "type": "integer"
          },
          "size": {
            "type": "integer"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "page",
          "size"
        ]
      },
...
      "rateHTTPResponse": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "rate": {
            "type": "integer"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "rate"
        ]
      },
...
      "sessionHistoryPageHTTPResponse": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "items": {
            "items": {
              "properties": {
                "sessionId": {
                  "type": "string"
                },
                "model": {
                  "type": "string"
                },
                "messageCount": {
                  "type": "integer"
                },
                "tokenCount": {
                  "type": "integer"
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time"
                }
              },
              "additionalProperties": false,
              "type": "object",
              "required": [
                "sessionId",
                "model",
                "messageCount",
                "tokenCount",
                "createdAt",
                "updatedAt"
              ]
            },
            "type": "array"
          },
          "total": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ]
      },
...
      "updateAuthUserRequest": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "firstName",
          "lastName"
        ]
      },
...
      "userStatsPageHTTPResponse": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "items": {
            "items": {
              "properties": {
                "email": {
                  "type": "string"
                },
                "firstName": {
                  "type": "string"
                },
                "lastName": {
                  "type": "string"
                },
                "totalChatMessages": {
                  "type": "integer"
                },
                "totalChatMessagesTokenCount": {
                  "type": "integer"
                },
                "totalChatMessages3Days": {
                  "type": "integer"
                },
                "totalChatMessages3DaysTokenCount": {
                  "type": "integer"
                },
                "avgChatMessages3DaysTokenCount": {
                  "type": "integer"
                },
                "rateLimit": {
                  "type": "integer"
                }
              },
              "additionalProperties": false,
              "type": "object",
              "required": [
                "email",
                "firstName",
                "lastName",
                "totalChatMessages",
                "totalChatMessagesTokenCount",
                "totalChatMessages3Days",
                "totalChatMessages3DaysTokenCount",
                "avgChatMessages3DaysTokenCount",
                "rateLimit"
              ]
            },
            "type": "array"
          },
          "total": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "Browser JWT access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  }
}